+++ __utils__/prepare-temp-dir/tsconfig.json
@@ -8,9 +8,5 @@
     "src/**/*.ts",
     "../../__typings__/**/*.d.ts"
   ],
-  "references": [
-    {
-      "path": "../prepare-temp-dir"
-    }
-  ]
+  "references": []
 }
