+++ __utils__/prepare/tsconfig.json
@@ -14,6 +14,9 @@
     },
     {
       "path": "../assert-project"
+    },
+    {
+      "path": "../prepare-temp-dir"
     }
   ]
 }
