+++ __utils__/test-fixtures/src/index.ts
@@ -1,6 +1,6 @@
 import fs from 'fs'
 import path from 'path'
-import { tempDir } from '@pnpm/prepare'
+import { tempDir } from '@pnpm/prepare-temp-dir'
 
 export interface FixturesHandle {
   copy: (name: string, dest: string) => void
